Update nws_alerts.py
I have added the configuration option "status_types" which will skip alerts that don't have a matching status type to those chosen in your configuration.  Some NWS counties constantly output alerts with the 'Test' status, which can cause nuisance notifications.  Configuration would look like this: 

- platform: nws_alerts
  zone_id: 'XXXXXX,XXXXXX'
  status_types: 
    - 'actual'
    - 'test'

I have incorporated severity in to the attributes.  The state value in my version reflects the severity "level", instead of the number of alerts.   It is a rank 0-4 from unknown to extreme.  I found this to be simpler to use for automation triggers.